requirements.txt
openai>=1.0.0
streamlit>=1.30.0
python-docx>=1.0.0
python-dotenv>=1.0.0
PyDrive>=1.3.1
google-generativeai>=0.3.2